Don't create dynamic classes for non-dynamic managers
We don't need to create custom managers when the manager isn't
dynamically created. This changes the logic to just set the type of the
manager on the class.

There's a small regression/limitations with this approach, as
demonstrated by the
custom_manager_without_typevar_returns_any_model_types test case. If the
manager is _not_ typed with generics (ie. `MyManager(Manager)`
vs `MyManager(Manager[T])` the manager will return querysets with `Any`
as the model type instead because of implicit generics.

That can be solved in one of two ways: either through
"reparametrization" as implemented in #1030 possibly or through setting
the methods as attributes with type `Any` as we do for dynamic managers