Remove dependency of AUTH_USER_MODEL for HttpRequest.
For modules "django.http" and "django.http.request", we manually add
dependency for the module containing "AUTH_USER_MODEL" for mypy to
resolve first.

This potentially adds cyclic dependency when the models module contains
reference to django.conf, and the settings depend on any modules that
rely on "django.http" or "django.http.request". At this point, the types
of the settings are only partially resolved.

Signed-off-by: Zixuan James Li <p359101898@gmail.com>