Fix type of Model field attributes on class (not instance)
For example:

```python
from django.contrib.auth.models import User
User.username
```

Returns `DeferredAttribute` instance, but it was previously  typehinted as `Self`.

Created a new `_FieldDescriptor` protocol that is compatible with all field descriptor implementations to use on the base `Field` class.

This was highlighted with mypy 1.3.0, it no longer considered the `__get__` overloads compatible, which is correct.