Annotate Migration fields as a ClassVar
For the fields "initial" and "atomic", this is strictly correct, as these
are set on the class, not the instance.

For "operations", "dependencies", "run_before", and "replaces", Django
intentionally copies the class attributes to the instance during
construction. However, annotating the attributes as either (using "|")
class or instance variables causes MyPy to issue the warning:
"Cannot override instance variable (previously declared on base
class "Migration") with class variable".
Since Django projects will nearly always override these as class
variables, annotate them as such.