Fix wrong type hints for SyndicationFeed methods (#1705)
Django's documentation mentions that the handler is of type
xml.sax.saxutils.XMLGenerator. However, it is actually of type
django.utils.xmlutils.SimplerXMLGenerator, which is a subclass of the
former.

Methods like `addQuickElement` (which are used in django's own
implementation of add_root_elements) is actually defined for
SimplerXMLGenerator, so this seems like the best fit, given that
subclasses will commonly rely on this method.