Add explicit overloads to each field class
This allows type checkers other than mypy to infer correctly
the `__set/get__` type variables, without the need for the plugin.

The plugin logic to solve these type variables is kept, as some
special casing is implemented to handle primary key fields with a default.
The `Field` docstring was updated with an example reflecting this