Relax type for `fields` argument of `Model.refresh_from_db()`
According to the docstring, any iterable or `None` is accepted.

There is a bug that is fixed in Django's `main` branch but not in
any release as of the time writing, where `remove()` can be called
on the `fields` argument.

https://code.djangoproject.com/ticket/34925

Since this is a bug, I think it should not affect the annotation.
In any case, `remove()` is only part of `MutableSequence`, not
`Sequence`, so the old annotation matched neither the documentation
nor the implementation.