Update `admin.sites._ViewType` bound to allow `StreamingHttpResponse`
From the [Django documentation](https://docs.djangoproject.com/en/5.0/ref/request-response/#streaminghttpresponse-objects):
> The `StreamingHttpResponse` is not a subclass of `HttpResponse` ...

Because of this, creating an admin view that returned a
`StreamingHttpResponse` was causing the following Mypy error:

```
testing/admin.py:14: error: Value of type variable "_ViewType" of "admin_view" of "AdminSite" cannot be "Callable[[HttpRequest], StreamingHttpResponse]"  [type-var]
```

This updates the admin `_ViewType` type-var bound to allow a Callable
that can return a union of `HttpResponse | StreamingHttpResponse`.

After this update to the stubs file, the Mypy error is no longer
reproducible.