Check correct model on other side of many to many reverse filtering
Depending on the manager type there could be different model arguments.
For a `ManyRelatedManager` the argument is the `through` model, but we
want to check the `to` model.

As a simple starter we stash the `to` model in the manager's metadata
and set a precedence on fetching the `to` model from metadata.