Resolve configured `AUTH_USER_MODEL` with a `get_type_analyze_hook`
It's nothing special really. We declare a `TypeAlias` in the pyi, thus
getting a fullname, somewhere under `django.contrib.auth`. We then build
a hook for that fullname. And the hook simulates what
`django.contrib.auth.get_user_model` does

The alias is set up to point to `AbstractBaseUser`, so for a type
checker other than mypy nothing should've changed