Allow non-string values in dummy template render context
The implementation of the `render()` method uses the standard library
method `string.Template.safe_substitute()` to do the actual formatting,
which accepts other value types than `str` and simply calls `__str__()`
on them.

The `safe_substitute()` method is annotated in typeshed with `object`
as the value type for the mapping. However, as `_EngineTemplate` uses
`Any` instead, I went with that.