fix: Add `context_data` attribute to _MonkeyPatchedW/ASGIResponse.
As detailed in Django's documentation [0], when a non-standard template backend is being used, the `context_data` attribute is populated instead of `context` when the test API call was successful. In reality, I've seen this to be a populated attribute regardless of the backend template engine being used. Regardless, this near hidden attribute is missing and should be included to help improve unit testing when validating the response data contains the expected data.