Fix typing of `Choices` with `__empty__`
When `__empty__` is defined, the `.choices` and `.values` properties
will add a value for the blank choice, so we expect that `None` should
be part of the return type for the value, e.g. `list[str | None]`
instead of `list[str]` for `.values` and for `.choices` we expect
`list[tuple[str | None, _StrOrPromise]` instead of
`list[tuple[str, `_StrOrPromise]]`. This is too dynamic for the stubs
and so requires some tweaks in the plugin.