Fixed typing of `Choices` for custom mixed-in type
It's easy enough to get decent types for `IntegerChoices` and
`TextChoices` in the stubs because of subclassing `IntEnum` and
`StrEnum`, as well as having a dedicated base enum type.

But for custom choices with mixed-in base types, it's not possible to
get sensible typing as enums cannot be generic. As such, the type of the
value has to be `Any`. In the plugin, we can find the base type and use
that instead.