Improve typing of choices (#2582)
When `__empty__` is defined, the `.choices` and `.values` properties
will add a value for the blank choice, so we expect that `None` should
be part of the return type for the value, e.g. `list[str | None]`
instead of `list[str]` for `.values` and for `.choices` we expect
`list[tuple[str | None, _StrOrPromise]` instead of
`list[tuple[str, `_StrOrPromise]]`. This is too dynamic for the stubs
and so requires some tweaks in the plugin.

* Fixed typing of `Choices` for custom mixed-in type

It's easy enough to get decent types for `IntegerChoices` and
`TextChoices` in the stubs because of subclassing `IntEnum` and
`StrEnum`, as well as having a dedicated base enum type.

But for custom choices with mixed-in base types, it's not possible to
get sensible typing as enums cannot be generic. As such, the type of the
value has to be `Any`. In the plugin, we can find the base type and use
that instead.