Fix type resolution for choices in plugin (1)
The resolution of the attribute type was too naive and didn't handle
where a module is imported and the choices type is accessed as an
attribute from that module.

Eliminates 11,087 errors in a large repository that were raised when
upgrading from django-stubs v5.1.3 to v5.2.0 which included the changes
in a3f2a558fc6c5544b3b46a6e53e6f266272236f7.

Co-authored-by: Stephen Moore <delfick@users.noreply.github.com>