Add non-lazy label handling for choices in plugin
Because of the change to use `django.utils.functional._StrOrPromise` for
labels to improve accuracy, it can be a little painful - especially in
a very large codebase where translations / lazy strings are not used.

This change keeps the stubs unchanged for correctness, e.g. in pyright
for where there is no alternative, but tweaks the mypy plugin to handle
introspecting the members to determine whether there are no lazy strings
such that the types can be narrowed.

Eliminates 1,166 errors in a large repository that were raised when
upgrading from django-stubs v5.1.3 to v5.2.0 which included the changes
in a3f2a558fc6c5544b3b46a6e53e6f266272236f7.