Fix type resolution for choices in plugin (5)
The resolution of the attribute type was too naive and didn't handle
where a choices type is decomposed to a union of enum literals when a
comparison to a member is made in a branch.

Eliminates 33 errors in a large repository that were raised when
upgrading from django-stubs v5.1.3 to v5.2.0 which included the changes
in a3f2a558fc6c5544b3b46a6e53e6f266272236f7.

Co-authored-by: Stephen Moore <delfick@users.noreply.github.com>