Fix type resolution for choices in plugin (6)
The resolution of the attribute type was too naive and didn't handle
where a choices type overrides a property, e.g. label, and makes use
of `super()`.

Eliminates 1 error in a large repository that was raised when upgrading
from django-stubs v5.1.3 to v5.2.0 which included the changes in
a3f2a558fc6c5544b3b46a6e53e6f266272236f7.