Suppress some errors in choices plugin
This case is a little complex as we have a mix of choices types. In
theory we could handle this, but it would require more rework to handle
multiple types. For now, just make no changes and return the original
type.

Eliminates 11 errors in a large repository that were raised when
upgrading from django-stubs v5.1.3 to v5.2.0 which included the changes
in a3f2a558fc6c5544b3b46a6e53e6f266272236f7.