Restore support for mypy v1.13
Currently mypy v1.13 is the minimum supported version, but the fix made
to the choices plugin code in #2646 made use of `TypeInfo.enum_members`
which was introduced in v1.14. This commit adds a backport of the
version introduced in v1.14 which can be removed when `django-stubs`
bumps its minimum version.