add mypy plugin support for Django's storage framework
This change correctly resolves
``django.core.files.storage.default_storage`` and
``django.contrib.staticfiles.storage.staticfiles_storage`` which are
configured in the project's Django settings.

When ``django.core.files.storage.storages``, which has a dict like
interface, is accessed with a literal value, then the type of the
appropriate storage is returned.