Remove use of `force_union_syntax` option
Now that django-stubs targets Python 3.10+, this option serves no
purpose. Support for PEP 604 in error messages has been available since
mypy v1.4 when using Python 3.10+ and the minimum supported version in
django-stubs is now mypy v1.13.

Note that there are still some cases where `Union[...]` is used, but
these are due to be changed to PEP 604 syntax in mypy v1.17.