fix: allow to query decimal fields with int (#2838)
* fix: allow to query decimal fields with int

Django allow to query decimal fields with int, and given there should be no precision loss, we should accept ints to query such fields.

Note: even though we do accept floats when setting a value in a decimal field, I opted for only accepting ints in this PR, as to minimize the impact of the change.

Fixes #2831.

Co-authored-by: Sebastian Kreft <911768+sk-@users.noreply.github.com>
Co-authored-by: sobolevn <mail@sobolevn.me>