Make `TemplateView` (and `TemplateResponseMixin`) generic over the response type
This allows us to specify that the default response type for template
views is a `TemplateResponse`.

N.B. The `LogoutView` may return a response that is not a template
response (i.e. a redirect), so we annotate it as returning a less
specific type. We have not been more specific about exactly which types
it returns because doing so might require it to also become generic to
support subclasses that return different specific response types.

Co-authored-by: Charlie Denton <charlie@meshy.co.uk>